+++ lib/achilles.rb
@@ -1,5 +1,6 @@
 require "achilles/version"
 require "importmap-rails"
+require "turbo-rails"
 require "achilles/engine"
 
 module Achilles