+++ test/dummy/config/importmap.rb
@@ -1,2 +1,3 @@
+pin "@hotwired/turbo-rails", to: "turbo.min.js"
 pin "application"
 pin "demo_counter_component"